inheritance_1_2_1_5

+++ src/by/epam/lab/Byn.java
@@ -42,7 +42,7 @@ public class Byn implements Comparable<Byn> {
         return this;
     }
 
-    public Byn mul(double d) {
+    public Byn mul(int d) {
         this.costOfObject *= d;
         return this;
     }
